Modify Identity element to simulate capture rate
In our pipeline, we read and parse raw video data which is then passed to the decoder. This will happen at a greater rate than the frames would be captured in real life. The modifications made in this commit allow the identity element to simulate a capture rate of 24fps.